requirements.txt
airbyte
duckdb
anthropic
streamlit
python-dotenv
pandas
